pyproject.toml
[tool.pytest.ini_options]
addopts = "-vvv --cov-report=term-missing --cov=apps --cov-branch"


[tool.coverage.report]
omit = [
    "app/__init__.py",
    "app/admin.py",
    "app/migrations/*"
]

[tool.poetry]
name = "blogs-post"
version = "0.1.0"
description = "API Restful baseada em DRF para armazenar postagens de blog."
authors = ["Hermano Lustosa <hermano_lustosa@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
PyJWT = "^2.8.0"
argon2 = "^0.1.10"
argon2-cffi = "^23.1.0"
shortuuid = "^1.0.13"
django-filter = "^24.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
pre-commit = "^3.7.0"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
